% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{apply_missingness}
\alias{apply_missingness}
\title{Apply missingness to a vector}
\usage{
apply_missingness(x, na_rate, mnar_score = 0)
}
\arguments{
\item{x}{Numeric vector}

\item{na_rate}{Proportion to make NA}

\item{mnar_score}{MNAR intensity (0 = MCAR)}
}
\value{
Vector with some values replaced by NA
}
\description{
Apply missingness to a vector
}
\keyword{internal}
