% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits.R
\name{avail_dists}
\alias{avail_dists}
\title{Available Distributions}
\usage{
avail_dists()
}
\value{
A character vector of available distribution names.
}
\description{
This provides a vector of available distributions for fitting.
}
