% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{bf_jzs}
\alias{bf_jzs}
\title{JZS Bayes factor approximation}
\usage{
bf_jzs(t, n1, n2)
}
\arguments{
\item{t}{t-statistic}

\item{n1}{Sample size group 1}

\item{n2}{Sample size group 2}
}
\value{
Bayes factor
}
\description{
JZS Bayes factor approximation
}
\keyword{internal}
