% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{compute_perpeptide_power}
\alias{compute_perpeptide_power}
\title{Compute per-peptide power (helper for plot_power_vs_effect)}
\usage{
compute_perpeptide_power(
  fits_data,
  n_per_group,
  effect_size,
  alpha,
  test,
  n_sim
)
}
\description{
Compute per-peptide power (helper for plot_power_vs_effect)
}
\keyword{internal}
