% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{compute_t_stat}
\alias{compute_t_stat}
\title{Compute t-statistic for two groups}
\usage{
compute_t_stat(x, y)
}
\arguments{
\item{x}{First group}

\item{y}{Second group}
}
\value{
t-statistic
}
\description{
Compute t-statistic for two groups
}
\keyword{internal}
