% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits.R
\name{d2n}
\alias{d2n}
\title{Distribution Name to Normalized Name}
\usage{
d2n(tag)
}
\arguments{
\item{tag}{A character string representing the distribution tag.}
}
\value{
A character string with the human-readable distribution name.
}
\description{
This converts distribution tags to names that humans can read and understand.
}
