% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits.R
\name{do_fits}
\alias{do_fits}
\title{Fit Multiple Distributions}
\usage{
do_fits(df)
}
\arguments{
\item{df}{A data frame with the data to fit.}
}
\value{
A data frame with the fit results for each distribution.
}
\description{
This is used to fits multiple distributions to the first column of a data frame.
}
