% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{evaldist}
\alias{evaldist}
\title{Evaluate Distributions}
\usage{
evaldist(fits_df, fit_col = "fits")
}
\arguments{
\item{fits_df}{A data frame containing the fit results.}

\item{fit_col}{A character string specifying the column name that contains the nested fit results. Default is "fits".}
}
\value{
A ggplot object showing the evaluation of fitted models.
}
\description{
This evaluates and plots the best-fitted and failed models from the fits data frame.
}
