% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{fake_leg}
\alias{fake_leg}
\title{Fake Legend}
\usage{
fake_leg(name = "Dist")
}
\arguments{
\item{name}{A character string specifying the name of the legend. Default is "Dist".}
}
\value{
A ggplot2 legend object.
}
\description{
This generates a fake legend for the distribution plots.
}
