% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{fill_scale}
\alias{fill_scale}
\title{Fill Scale for Distributions}
\usage{
fill_scale(name = "name")
}
\arguments{
\item{name}{A character string specifying the name of the fill scale.}
}
\value{
A ggplot2 scale fill manual object.
}
\description{
This generates a fill scale for distributions using the Set3 palette from RColorBrewer.
}
