% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{find_effect_size}
\alias{find_effect_size}
\title{Find minimum detectable effect size}
\usage{
find_effect_size(
  distribution,
  params,
  n_per_group,
  target_power,
  alpha,
  test,
  n_sim
)
}
\arguments{
\item{distribution}{Distribution name}

\item{params}{Distribution parameters}

\item{n_per_group}{Sample size per group}

\item{target_power}{Target power level}

\item{alpha}{Significance level}

\item{test}{Statistical test}

\item{n_sim}{Number of simulations per effect size}
}
\value{
List with effect_size and effect_curve
}
\description{
Find minimum detectable effect size
}
\keyword{internal}
