% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peppwr.R
\name{find_fits}
\alias{find_fits}
\title{Find Fits for Nested Data}
\usage{
find_fits(df, id_col = "id", group_col = "group", value_col = "value")
}
\arguments{
\item{df}{A data frame containing the data to be fitted.}

\item{id_col}{A character string specifying the column name for the identifier. Default is "id".}

\item{group_col}{A character string specifying the column name for the group. Default is "group".}

\item{value_col}{A character string specifying the column name for the values to be fitted. Default is "value".}
}
\value{
A data frame with nested data and the fit results for each group.
}
\description{
This groups the data frame by specified columns, nests the data, and applies the fitting functions to each group.
}
