% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{find_sample_size}
\alias{find_sample_size}
\title{Find required sample size for target power}
\usage{
find_sample_size(
  distribution,
  params,
  effect_size,
  target_power,
  alpha,
  test,
  n_sim
)
}
\arguments{
\item{distribution}{Distribution name}

\item{params}{Distribution parameters}

\item{effect_size}{Effect size to detect}

\item{target_power}{Target power level}

\item{alpha}{Significance level}

\item{test}{Statistical test}

\item{n_sim}{Number of simulations per sample size}
}
\value{
List with n and power_curve
}
\description{
Find required sample size for target power
}
\keyword{internal}
