% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{format_answer}
\alias{format_answer}
\title{Format answer based on question type}
\usage{
format_answer(answer, question)
}
\arguments{
\item{answer}{The answer value}

\item{question}{The question type}
}
\value{
Formatted string
}
\description{
Format answer based on question type
}
\keyword{internal}
