% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_dfunc}
\alias{get_dfunc}
\title{Get density function for a distribution}
\usage{
get_dfunc(distribution)
}
\arguments{
\item{distribution}{Distribution name}
}
\value{
Density function
}
\description{
Get density function for a distribution
}
\keyword{internal}
