% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{get_dist_params}
\alias{get_dist_params}
\title{Get distribution parameters from fitted results}
\usage{
get_dist_params(dist_name, fit_df, raw_data)
}
\arguments{
\item{dist_name}{Distribution name}

\item{fit_df}{Fit results data frame}

\item{raw_data}{Raw data values}
}
\value{
List of parameters
}
\description{
Get distribution parameters from fitted results
}
\keyword{internal}
