% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits.R
\name{get_distribution_preset}
\alias{get_distribution_preset}
\title{Get distributions for a preset}
\usage{
get_distribution_preset(preset = "continuous")
}
\arguments{
\item{preset}{One of "continuous" (default), "counts", or "all"}
}
\value{
Character vector of distribution names
}
\description{
Get distributions for a preset
}
