% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_qfunc}
\alias{get_qfunc}
\title{Get quantile function for a distribution}
\usage{
get_qfunc(distribution)
}
\arguments{
\item{distribution}{Distribution name}
}
\value{
Quantile function
}
\description{
Get quantile function for a distribution
}
\keyword{internal}
