% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{get_rfunc}
\alias{get_rfunc}
\title{Get the random generation function for a distribution}
\usage{
get_rfunc(distribution)
}
\arguments{
\item{distribution}{Distribution name}
}
\value{
Random generation function
}
\description{
Get the random generation function for a distribution
}
\keyword{internal}
