% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{get_test_func}
\alias{get_test_func}
\title{Get the test function for a given test name}
\usage{
get_test_func(test)
}
\arguments{
\item{test}{Test name}
}
\value{
Test function
}
\description{
Get the test function for a given test name
}
\keyword{internal}
