% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits.R
\name{interpret_dataset_mnar}
\alias{interpret_dataset_mnar}
\title{Interpret dataset-level MNAR result}
\usage{
interpret_dataset_mnar(correlation, p_value, n_peptides)
}
\arguments{
\item{correlation}{Spearman correlation coefficient}

\item{p_value}{p-value for the correlation}

\item{n_peptides}{Number of peptides in calculation}
}
\value{
Interpretation string
}
\description{
Interpret dataset-level MNAR result
}
\keyword{internal}
