% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{map_dist_name}
\alias{map_dist_name}
\title{Map distribution name to R function prefix}
\usage{
map_dist_name(dist_name)
}
\arguments{
\item{dist_name}{Distribution name}
}
\value{
R function name for random generation
}
\description{
Map distribution name to R function prefix
}
\keyword{internal}
