% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_peppwr_fits}
\alias{new_peppwr_fits}
\title{Create a new peppwr_fits object}
\usage{
new_peppwr_fits(data, fits, best, call)
}
\arguments{
\item{data}{Original data (nested tibble)}

\item{fits}{Fit results per peptide (list of tibbles with dist, loglik, aic)}

\item{best}{Best-fitting distribution per peptide (character vector)}

\item{call}{Original function call}
}
\value{
A peppwr_fits object
}
\description{
Create a new peppwr_fits object
}
