% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_peppwr_power}
\alias{new_peppwr_power}
\title{Create a new peppwr_power object}
\usage{
new_peppwr_power(mode, question, answer, simulations, params, call)
}
\arguments{
\item{mode}{Either "aggregate" or "per_peptide"}

\item{question}{What was solved for: "power", "sample_size", or "effect_size"}

\item{answer}{The computed answer}

\item{simulations}{List of simulation details}

\item{params}{List of input parameters used}

\item{call}{Original function call}
}
\value{
A peppwr_power object
}
\description{
Create a new peppwr_power object
}
