% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits.R
\name{parse_fitdist}
\alias{parse_fitdist}
\title{Parse Fit Results from fitdistrplus}
\usage{
parse_fitdist(fit)
}
\arguments{
\item{fit}{An object returned by \code{fitdistrplus::fitdist}.}
}
\value{
A tibble with the distribution name, log-likelihood, and AIC.
}
\description{
This converts fit results from the \code{fitdistrplus} package into a tidy format.
}
