% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits.R
\name{parse_univariateML}
\alias{parse_univariateML}
\title{Parse Fit Results from univariateML}
\usage{
parse_univariateML(fit)
}
\arguments{
\item{fit}{An object returned by \code{univariateML::model_select}.}
}
\value{
A tibble with the distribution name, log-likelihood, and AIC.
}
\description{
This converts fit results from the \code{univariateML} package into a tidy format.
}
