% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.peppwr_fits}
\alias{plot.peppwr_fits}
\title{Plot method for peppwr_fits}
\usage{
\method{plot}{peppwr_fits}(x, ...)
}
\arguments{
\item{x}{A peppwr_fits object}

\item{...}{Additional arguments (ignored)}
}
\value{
A ggplot object
}
\description{
Creates a bar chart showing the count of best-fit distributions
}
