% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.peppwr_power}
\alias{plot.peppwr_power}
\title{Plot method for peppwr_power}
\usage{
\method{plot}{peppwr_power}(x, ...)
}
\arguments{
\item{x}{A peppwr_power object}

\item{...}{Additional arguments (ignored)}
}
\value{
A ggplot object
}
\description{
Creates power curves or \% peptides at threshold plots
}
