% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_best}
\alias{plot_best}
\title{Plot Best Fitted Models}
\usage{
plot_best(fits_df, fit_col = "fits")
}
\arguments{
\item{fits_df}{A data frame containing the fit results.}

\item{fit_col}{A character string specifying the column name that contains the nested fit results. Default is "fits".}
}
\value{
A ggplot object showing the best-fitted models.
}
\description{
This plots the best-fitted models based on Log-Likelihood and AIC from the fits data frame.
}
