% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_failed}
\alias{plot_failed}
\title{Plot Failed Models}
\usage{
plot_failed(fits_df, fit_col = "fits")
}
\arguments{
\item{fits_df}{A data frame containing the fit results.}

\item{fit_col}{A character string specifying the column name that contains the nested fit results. Default is "fits".}
}
\value{
A ggplot object showing the count of failed models.
}
\description{
This plots the count of failed model fits from the fits data frame.
}
