% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_missingness}
\alias{plot_missingness}
\title{Plot missingness statistics}
\usage{
plot_missingness(fits)
}
\arguments{
\item{fits}{A peppwr_fits object}
}
\value{
A ggplot object
}
\description{
Shows distribution of NA rates and MNAR scores across peptides.
}
