% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_param_distribution}
\alias{plot_param_distribution}
\title{Plot distribution of fitted parameters across peptidome}
\usage{
plot_param_distribution(fits)
}
\arguments{
\item{fits}{A peppwr_fits object}
}
\value{
A ggplot object
}
\description{
Plot distribution of fitted parameters across peptidome
}
