% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_power_aggregate}
\alias{plot_power_aggregate}
\title{Plot power curve for aggregate mode}
\usage{
plot_power_aggregate(x)
}
\arguments{
\item{x}{A peppwr_power object}
}
\value{
A ggplot object
}
\description{
Plot power curve for aggregate mode
}
\keyword{internal}
