% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_power_heatmap}
\alias{plot_power_heatmap}
\title{Plot power heatmap: N x effect size grid}
\usage{
plot_power_heatmap(
  distribution,
  params,
  n_range,
  effect_range,
  n_steps = 6,
  n_sim = 100,
  test = "wilcoxon",
  alpha = 0.05
)
}
\arguments{
\item{distribution}{Distribution name}

\item{params}{Distribution parameters}

\item{n_range}{Range of sample sizes (vector of length 2)}

\item{effect_range}{Range of effect sizes (vector of length 2)}

\item{n_steps}{Number of grid points per dimension}

\item{n_sim}{Number of simulations per grid cell}

\item{test}{Statistical test to use}

\item{alpha}{Significance level}
}
\value{
A ggplot object
}
\description{
Plot power heatmap: N x effect size grid
}
