% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_power_perpeptide}
\alias{plot_power_perpeptide}
\title{Plot \% peptides at threshold for per-peptide mode}
\usage{
plot_power_perpeptide(x)
}
\arguments{
\item{x}{A peppwr_power object}
}
\value{
A ggplot object
}
\description{
Plot \% peptides at threshold for per-peptide mode
}
\keyword{internal}
