% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_power_vs_effect}
\alias{plot_power_vs_effect}
\title{Plot power vs effect size at fixed N}
\usage{
plot_power_vs_effect(power_result, effect_range, n_steps = 10, n_sim = NULL)
}
\arguments{
\item{power_result}{A peppwr_power object}

\item{effect_range}{Range of effect sizes to explore}

\item{n_steps}{Number of effect size values to compute}

\item{n_sim}{Number of simulations per point}
}
\value{
A ggplot object
}
\description{
Plot power vs effect size at fixed N
}
