% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_qq}
\alias{plot_qq}
\title{Plot QQ plots for goodness-of-fit}
\usage{
plot_qq(fits, peptide_id = NULL, n_plots = 6)
}
\arguments{
\item{fits}{A peppwr_fits object}

\item{peptide_id}{Specific peptide to plot (NULL for multiple)}

\item{n_plots}{Number of peptides to plot when peptide_id is NULL}
}
\value{
A ggplot object
}
\description{
Plot QQ plots for goodness-of-fit
}
