% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power_analysis}
\alias{power_analysis}
\title{Power analysis for peptide abundance data}
\usage{
power_analysis(distribution, ...)
}
\arguments{
\item{distribution}{Distribution name (character) or peppwr_fits object for per-peptide mode}

\item{...}{Additional arguments}
}
\value{
A peppwr_power object
}
\description{
Power analysis for peptide abundance data
}
