% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.peppwr_fits}
\alias{print.peppwr_fits}
\title{Print method for peppwr_fits}
\usage{
\method{print}{peppwr_fits}(x, ...)
}
\arguments{
\item{x}{A peppwr_fits object}

\item{...}{Additional arguments (ignored)}
}
\value{
The object invisibly
}
\description{
Print method for peppwr_fits
}
