% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.peppwr_power}
\alias{print.peppwr_power}
\title{Print method for peppwr_power}
\usage{
\method{print}{peppwr_power}(x, ...)
}
\arguments{
\item{x}{A peppwr_power object}

\item{...}{Additional arguments (ignored)}
}
\value{
The object invisibly
}
\description{
Print method for peppwr_power
}
