% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.summary.peppwr_fits}
\alias{print.summary.peppwr_fits}
\title{Print method for summary.peppwr_fits}
\usage{
\method{print}{summary.peppwr_fits}(x, ...)
}
\arguments{
\item{x}{A summary.peppwr_fits object}

\item{...}{Additional arguments (ignored)}
}
\value{
The object invisibly
}
\description{
Print method for summary.peppwr_fits
}
