% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{run_power_sim}
\alias{run_power_sim}
\title{Run power simulation}
\usage{
run_power_sim(
  distribution,
  params,
  n_per_group,
  effect_size,
  alpha = 0.05,
  test = "wilcoxon",
  n_sim = 1000
)
}
\arguments{
\item{distribution}{Distribution name}

\item{params}{List of distribution parameters}

\item{n_per_group}{Number of samples per group}

\item{effect_size}{Fold change for treatment}

\item{alpha}{Significance level}

\item{test}{Statistical test to use ("wilcoxon", "bootstrap_t", "bayes_t", "rankprod")}

\item{n_sim}{Number of simulations}
}
\value{
Power estimate (proportion of significant results)
}
\description{
Run power simulation
}
