% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{run_power_sim_with_missingness}
\alias{run_power_sim_with_missingness}
\title{Run power simulation with missingness}
\usage{
run_power_sim_with_missingness(
  distribution,
  params,
  n_per_group,
  effect_size,
  na_rate = 0,
  mnar_score = 0,
  alpha = 0.05,
  test = "wilcoxon",
  n_sim = 1000
)
}
\arguments{
\item{distribution}{Distribution name}

\item{params}{List of distribution parameters}

\item{n_per_group}{Number of samples per group}

\item{effect_size}{Fold change for treatment}

\item{na_rate}{Proportion of values that will be NA}

\item{mnar_score}{MNAR intensity (0 = MCAR)}

\item{alpha}{Significance level}

\item{test}{Statistical test to use}

\item{n_sim}{Number of simulations}
}
\value{
Power estimate (proportion of significant results)
}
\description{
Estimates power accounting for realistic missing data patterns.
}
