% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits.R
\name{single_fit}
\alias{single_fit}
\title{Fit a Single Distribution}
\usage{
single_fit(df, dist)
}
\arguments{
\item{df}{The data frame containing the data to fit.}

\item{dist}{The distribution name or a list of distribution models to try fitting. One of "nbinom", "gamma", "snorm", "norm",
"invgamma", "invgauss", "lnorm", "lgamma",  and "pareto"}
}
\value{
A list containing the fit result or the distribution name in case there is an error.
}
\description{
This fits a specified distribution to the first column of a data frame.
}
