% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits.R
\name{squash_fits}
\alias{squash_fits}
\title{Squash Fit Results}
\usage{
squash_fits(fit)
}
\arguments{
\item{fit}{A list containing the fit result.}
}
\value{
A tibble with the distribution name, log-likelihood, and Akaike Information Criterion (AIC).
}
\description{
This processes fit results into a tidy format.
}
