% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{summary.peppwr_fits}
\alias{summary.peppwr_fits}
\title{Summary method for peppwr_fits}
\usage{
\method{summary}{peppwr_fits}(object, ...)
}
\arguments{
\item{object}{A peppwr_fits object}

\item{...}{Additional arguments (ignored)}
}
\value{
A list with summary statistics
}
\description{
Summary method for peppwr_fits
}
