% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{summary.peppwr_power}
\alias{summary.peppwr_power}
\title{Summary method for peppwr_power}
\usage{
\method{summary}{peppwr_power}(object, ...)
}
\arguments{
\item{object}{A peppwr_power object}

\item{...}{Additional arguments (ignored)}
}
\value{
A list with summary statistics
}
\description{
Summary method for peppwr_power
}
