% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test_bayes_t}
\alias{test_bayes_t}
\title{Bayes factor t-test}
\usage{
test_bayes_t(control, treatment)
}
\arguments{
\item{control}{Control group values}

\item{treatment}{Treatment group values}
}
\value{
Bayes factor (BF10: evidence for alternative over null)
}
\description{
Computes Bayes factor for difference between two groups
}
