% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test_bootstrap_t}
\alias{test_bootstrap_t}
\title{Bootstrap-t test}
\usage{
test_bootstrap_t(control, treatment, n_boot = 1000)
}
\arguments{
\item{control}{Control group values}

\item{treatment}{Treatment group values}

\item{n_boot}{Number of bootstrap iterations (default 1000)}
}
\value{
p-value from the test
}
\description{
Performs a bootstrap-t test comparing two groups
}
