% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{test_wilcoxon}
\alias{test_wilcoxon}
\title{Wilcoxon rank-sum test}
\usage{
test_wilcoxon(control, treatment)
}
\arguments{
\item{control}{Control group values}

\item{treatment}{Treatment group values}
}
\value{
p-value from the test
}
\description{
Wilcoxon rank-sum test
}
