% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{validate_peppwr_fits}
\alias{validate_peppwr_fits}
\title{Validate a peppwr_fits object}
\usage{
validate_peppwr_fits(x)
}
\arguments{
\item{x}{Object to validate}
}
\value{
The validated object (invisibly)
}
\description{
Validate a peppwr_fits object
}
