% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{validate_peppwr_power}
\alias{validate_peppwr_power}
\title{Validate a peppwr_power object}
\usage{
validate_peppwr_power(x)
}
\arguments{
\item{x}{Object to validate}
}
\value{
The validated object (invisibly)
}
\description{
Validate a peppwr_power object
}
